방명록 구현 (페이징 x)
페이징하다 때려침 언젠가 건드려야함.

+++ src/main/java/org/interior/repository/VisitorRepository.java
@@ -1,7 +1,8 @@
 package org.interior.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 								//<테이블이름, id의 자료형타입>
-public interface VisitorRepository extends JpaRepository<Visitor, Long>{
+public interface VisitorRepository extends JpaRepository<Visitor, Integer>{
 	
 }
